Selects going through
Okay so selects go through now. We still have some big jank there and on
top of that, the current implementation of messaging breaks when we
change tabs to another, as the queries don't really go through. That
means we have to find out what's causing the tab call mismatch