One step forwards, 5 steps back
So the thing is, we don't want to analyze too much ourselves. This came
clear from the TS / CEM trials... After that, I've started writing the
basic layer of property/attribute accessors, which also cause a little
bit of a headache as they need to have info on what was selected, and
passing data between the Content Script and the Actual DOM can be
tricky.